build: restructure project files for building
