+++ .prettierignore
@@ -1,3 +1,2 @@
-dist/
-pipx-install-action/
+dist
 pnpm-lock.yaml
