+++ action.yml
@@ -10,4 +10,4 @@ inputs:
     required: true
 runs:
   using: node20
-  main: dist/main.mjs
+  main: dist/action/main.bundle.mjs
