+++ package.json
@@ -18,11 +18,10 @@
   "license": "MIT",
   "author": "Alfi Maulana <alfi.maulana.f@gmail.com>",
   "type": "module",
-  "main": "dist/lib.js",
-  "types": "dist/lib.d.ts",
+  "main": "dist/lib/pipx.js",
+  "types": "dist/lib/pipx.d.ts",
   "files": [
-    "dist/lib.js",
-    "dist/lib.d.ts"
+    "dist/lib"
   ],
   "scripts": {
     "prepack": "rollup -c",
@@ -37,7 +36,8 @@
   "devDependencies": {
     "@eslint/js": "^9.35.0",
     "@rollup/plugin-node-resolve": "^16.0.1",
-    "@tsconfig/node23": "^23.0.2",
+    "@rollup/plugin-typescript": "^12.1.4",
+    "@tsconfig/node20": "^20.1.6",
     "@types/jest": "^30.0.0",
     "@types/node": "^24.3.0",
     "@vitest/coverage-v8": "^3.0.8",
@@ -46,7 +46,7 @@
     "prettier": "^3.6.2",
     "prettier-plugin-organize-imports": "^4.2.0",
     "rollup": "^4.50.0",
-    "rollup-plugin-ts": "^3.4.5",
+    "tslib": "^2.8.1",
     "typedoc": "^0.28.12",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.42.0",
