+++ rollup.config.js
@@ -1,33 +1,11 @@
 import { nodeResolve } from "@rollup/plugin-node-resolve";
-import ts from "rollup-plugin-ts";
+import typescript from "@rollup/plugin-typescript";
 
-export default [
-  {
-    input: "src/lib.ts",
-    output: {
-      dir: "dist",
-    },
-    plugins: [
-      ts({
-        tsconfig: (config) => ({ ...config, declaration: true }),
-        transpileOnly: true,
-      }),
-    ],
-    external: [
-      "cache-action",
-      "catched-error-message",
-      "gha-utils",
-      "node:child_process",
-      "os",
-      "path",
-    ],
+export default {
+  input: "src/action/main.ts",
+  output: {
+    dir: "dist/action",
+    entryFileNames: "[name].bundle.mjs",
   },
-  {
-    input: "src/main.ts",
-    output: {
-      dir: "dist",
-      entryFileNames: "[name].mjs",
-    },
-    plugins: [nodeResolve(), ts({ transpileOnly: true })],
-  },
-];
+  plugins: [nodeResolve(), typescript()],
+};
