+++ src/lib/pipx/cache.test.ts
