+++ src/lib/pipx/environment.test.ts
