+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
-  "extends": "@tsconfig/node23",
+  "extends": "@tsconfig/node20",
   "compilerOptions": {
+    "module": "node16",
     "noEmit": true
   }
 }
