+++ typedoc.json
@@ -1,3 +1,3 @@
 {
-  "entryPoints": ["src/lib.ts"]
+  "entryPoints": ["src/lib/pipx.ts"]
 }
